test:add ManagerTest and BlockStore unit test
