+++ src/test/java/org/tron/core/db/BlockStoreTest.java
@@ -6,10 +6,8 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.tron.core.Constant;
-import org.tron.core.capsule.TransactionCapsule;
 import org.tron.core.config.Configuration;
 import org.tron.core.config.args.Args;
-import org.tron.core.exception.ValidateException;
 
 
 public class BlockStoreTest {
@@ -19,21 +17,17 @@ public class BlockStoreTest {
   @Before
   public void init() {
     Args.setParam(new String[]{}, Configuration.getByPath(Constant.TEST_CONF));
-
   }
 
   @Test
-  public void testPushTransactions() {
-    Args.setParam(new String[]{}, Configuration.getByPath(Constant.TEST_CONF));
-    BlockStore blockStore = BlockStore.create("test_OperateTx");
-    TransactionCapsule transactionCapsule = new TransactionCapsule(
-        "2c0937534dd1b3832d05d865e8e6f2bf23218300b33a992740d45ccab7d4f519", 123);
-    try {
-      blockStore.pushTransactions(transactionCapsule);
-    } catch (ValidateException e) {
-      e.printStackTrace();
-    }
-    Assert
-        .assertEquals(123, transactionCapsule.getTransaction().getRawData().getVout(0).getValue());
+  public void testCreateBlockStore() {
+    BlockStore blockStore = BlockStore.create("test-createBlock");
+    Assert.assertEquals("0000000000000000000000000000000000000000000000000000000000000000",
+        blockStore.getHeadBlockId().toString());
+    Assert.assertEquals(0, blockStore.getHeadBlockNum());
+    Assert.assertEquals(blockStore.getGenesisTime(), blockStore.getHeadBlockTime());
+    Assert.assertEquals(0, blockStore.currentASlot());
+    blockStore.close();
+
   }
 }
