+++ src/test/java/org/tron/core/db/ManagerTest.java
@@ -2,7 +2,7 @@ package org.tron.core.db;
 
 import com.google.protobuf.ByteString;
 import org.junit.Assert;
-import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,24 +11,25 @@ import org.tron.common.utils.ByteArray;
 import org.tron.core.Constant;
 import org.tron.core.Sha256Hash;
 import org.tron.core.capsule.BlockCapsule;
+import org.tron.core.capsule.TransactionCapsule;
 import org.tron.core.config.Configuration;
 import org.tron.core.config.args.Args;
+import org.tron.core.exception.ValidateException;
 
 public class ManagerTest {
 
   private static final Logger logger = LoggerFactory.getLogger("Test");
+  static Manager dbManager = new Manager();
 
-  @Before
-  public void init() {
+  @BeforeClass
+  public static void init() {
     Args.setParam(new String[]{}, Configuration.getByPath(Constant.TEST_CONF));
-
+    dbManager.init();
   }
 
   @Test
   public void pushBlock() {
 
-    Manager dbManager = new Manager();
-    dbManager.init();
     BlockCapsule blockCapsule2 = new BlockCapsule(0, ByteString.copyFrom(ByteArray
         .fromHexString("0304f784e4e7bae517bcab94c3e0c9214fb4ac7ff9d7d5a937d1f40031f87b81")),
         0,
@@ -37,7 +38,11 @@ public class ManagerTest {
                 .getAddress()));
     blockCapsule2.setMerklerRoot();
     blockCapsule2.sign(Args.getInstance().getPrivateKey().getBytes());
-    dbManager.pushBlock(blockCapsule2);
+    try {
+      dbManager.pushBlock(blockCapsule2);
+    } catch (ValidateException e) {
+      e.printStackTrace();
+    }
 
     Assert.assertTrue(dbManager.containBlock(Sha256Hash.wrap(ByteArray
         .fromHexString("c37fea1dec8048180911c6cf075348f93a524336c47e97317eb59b91bd485ca4"))));
@@ -57,4 +62,16 @@ public class ManagerTest {
         .fromHexString("c37fea1dec8048180911c6cf075348f93a524336c47e97317eb59b91bd485ca4"))));
   }
 
+  @Test
+  public void testPushTransactions() {
+    TransactionCapsule transactionCapsule = new TransactionCapsule(
+        "2c0937534dd1b3832d05d865e8e6f2bf23218300b33a992740d45ccab7d4f519", 123);
+    try {
+      dbManager.pushTransactions(transactionCapsule);
+    } catch (ValidateException e) {
+      e.printStackTrace();
+    }
+    Assert
+        .assertEquals(123, transactionCapsule.getTransaction().getRawData().getVout(0).getValue());
+  }
 }
